[MARKET-317] Force IE document mode to edge
The embedded IE10 browser (in Windows 8) was falling into IE9 document mode
(even thou it rendered in IE10 document mode when directly in the browser).

This enables the text input data binding when pressing backspace, but only
in IE10 and latter and not totally stable.

IE9 (and earlier) would need a fix avaliable only in angular.js 1.4.x branch,
but we are using angular.js 1.2.x branch for IE8 compatibility.